requirements.txt
datasets
dateutils
deepdiff
diffusers
dreamsim
einops
fastapi
h5py
h5pickle
huggingface-hub
hydra-core
lightning
matplotlib
multiprocess
numpy
omegaconf
pandas
Pillow
plotly
pyarrow
pytorch-fid
pytorch-lightning
rootutils
scikit-learn
scipy
seaborn
sentencepiece
sentry-sdk
skia-python
tables
tensorboard
tensorboard-data-server
timm
tokenizers
torch
torchcde
torchcfm
torchdata
torchdiffeq
torchdyn
torchmetrics
torchsde
torchvision
tqdm
trampoline
transformers
wandb
Werkzeug